WIP: animation: ensure repaints are scheduled after animation frames
If a view is offscreen it's output_mask is 0. If an animation brings
the view onscreen the output_mask isn't updated before the output
thinks there's nothing else to draw, so stops updating. Once something
else forces the output to repaint (a move of the pointer, for example)
the original view's output_mask is updated and the animation comes to
a swift end, having lost lots of time in the middle.

I don't properly understand this problem yet and this is probably just
a hack to make things work. This should be fixed nicely (and upstream)
soon.